requirements.txt
anyio==4.4.0
certifi==2024.2.2
elastic-transport==8.13.1
elasticsearch==8.13.2
elasticsearch-dsl==8.13.1
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Levenshtein==0.25.1
mysql-connector-python==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-telegram-bot==21.2
rapidfuzz==3.9.3
six==1.16.0
sniffio==1.3.1
urllib3==2.2.1
